Refactored whole project to allow npm publishing
- Renamed package as ra-data-opencrud
- Extracted backoffice into /examples folder
- Extracted actual library into /src folder
- Added some babel dependencies to transpile some ES6 features
