+++ .babelrc
@@ -1,0 +1,8 @@
+{
+  "plugins": ["transform-es2015-destructuring", "transform-object-rest-spread"],
+  "env": {
+    "test": {
+      "plugins": ["transform-es2015-modules-commonjs"]
+    }
+  }
+}
