+++ .gitignore
@@ -1,13 +1,15 @@
 # See https://help.github.com/ignore-files/ for more about ignoring files.
 
 # dependencies
-/node_modules
+node_modules/
 
 # testing
 /coverage
 
 # production
 /build
+/lib
+/esm
 
 # misc
 .DS_Store
