+++ examples/prisma-ecommerce/public/favicon.ico
