+++ examples/prisma-ecommerce/public/manifest.json
