+++ examples/prisma-ecommerce/src/components/attributes/index.js
