+++ examples/prisma-ecommerce/src/components/categories/index.js
