+++ examples/prisma-ecommerce/src/components/products/index.js
