+++ examples/prisma-ecommerce/src/components/shops/index.js
