+++ examples/prisma-ecommerce/src/index.css
