+++ examples/prisma-ecommerce/src/registerServiceWorker.js
