+++ package.json
@@ -1,31 +1,54 @@
 {
-  "name": "ra-data-prisma",
-  "version": "0.1.0",
-  "private": true,
+  "name": "ra-data-opencrud",
+  "version": "0.0.2",
+  "description": "A Prisma/GraphCMS data provider for react-admin",
+  "main": "lib/index.js",
+  "module": "esm/index.js",
+  "sideEffects": false,
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/weakky/ra-data-opencrud.git"
+  },
+  "bugs": {
+    "url": "https://github.com/weakky/ra-data-opencrud/issues"
+  },
+  "homepage": "https://github.com/weakky/ra-data-opencrud#readme",
+  "authors": [
+    "Flavian DESVERNE"
+  ],
+  "keywords": [
+    "reactjs",
+    "react",
+    "react-admin",
+    "admin-on-rest",
+    "rest",
+    "graphql",
+    "prisma",
+    "graphcms"
+  ],
+  "license": "MIT",
+  "scripts": {
+    "build": "rimraf ./lib && cross-env NODE_ENV=production BABEL_ENV=cjs babel --quiet ./src -d ./lib --ignore spec.js,test.js",
+    "build-esm": "rimraf ./esm && cross-env NODE_ENV=production BABEL_ENV=esm babel --quiet ./src -d ./esm --ignore spec.js,test.js",
+    "watch": "rimraf ./lib && cross-env NODE_ENV=production babel --watch ./src -d ./lib --ignore spec.js,test.js"
+  },
   "dependencies": {
-    "apollo-client": "^2.3.8",
-    "apollo-client-preset": "^1.0.8",
     "graphql-tag": "^2.6.1",
     "lodash": "~4.17.5",
     "pluralize": "~7.0.0",
-    "ra-data-graphql": "^2.3.0",
-    "react": "^16.4.1",
-    "react-admin": "^2.3.0",
-    "react-dom": "^16.4.1",
-    "react-scripts": "1.1.4"
+    "ra-data-graphql": "^2.3.0"
   },
   "peerDependencies": {
     "graphql": "^0.13.2",
     "react-admin": "^2.3.0"
   },
   "devDependencies": {
+    "babel-cli": "~6.26.0",
+    "babel-plugin-transform-es2015-destructuring": "^6.23.0",
+    "babel-plugin-transform-object-rest-spread": "^6.23.0",
+    "cross-env": "^5.2.0",
     "graphql": "^0.13.2",
-    "react-admin": "^2.3.0"
-  },
-  "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test --env=jsdom",
-    "eject": "react-scripts eject"
+    "react-admin": "^2.3.0",
+    "rimraf": "^2.6.2"
   }
 }
