+++ src/buildVariables.js
