+++ src/utils/getFinalType.test.js
