+++ src/utils/isList.js
